Login com nivel de acesso

+++ view/consulta_usuario.php
@@ -47,16 +47,31 @@
                             {
                                 $value->acesso = $value->acesso == 1 ? "Administrador" : "Usuário";                            
 
-                                echo "<tr>
-                                    <th scope='row'>$value->codusuario</th>
-                                    <td>$value->nome</td>
-                                    <td>$value->email</td>
-                                    <td>$value->acesso</td>
-                                    <td>
-                                        <a href='". URL ."excluir-usuario/$value->codusuario' onclick='return confirm(\"Tem certeza?\")' class='btn btn-outline-danger btn-sm'><i class='fa fa-trash'></i> Excluir</a>
-                                        <a href='". URL ."editar-usuario/$value->codusuario' class='btn btn-outline-warning btn-sm'><i class='fa fa-edit'></i> Editar</a>
-                                    </td>
-                                </tr>";
+                                if($_SESSION["sessao"]->acesso == 1)
+                                {
+                                    echo "<tr>
+                                        <th scope='row'>$value->codusuario</th>
+                                        <td>$value->nome</td>
+                                        <td>$value->email</td>
+                                        <td>$value->acesso</td>
+                                        <td>
+                                            <a href='". URL ."excluir-usuario/$value->codusuario' onclick='return confirm(\"Tem certeza?\")' class='btn btn-outline-danger btn-sm'><i class='fa fa-trash'></i> Excluir</a>
+                                            <a href='". URL ."editar-usuario/$value->codusuario' class='btn btn-outline-warning btn-sm'><i class='fa fa-edit'></i> Editar</a>
+                                        </td>
+                                    </tr>";
+                                }
+                                else
+                                {
+                                    echo "<tr>
+                                        <th scope='row'>$value->codusuario</th>
+                                        <td>$value->nome</td>
+                                        <td>$value->email</td>
+                                        <td>$value->acesso</td>
+                                        <td>
+                                            <a href='". URL ."editar-usuario/$value->codusuario' class='btn btn-outline-warning btn-sm'><i class='fa fa-edit'></i> Editar</a>
+                                        </td>
+                                    </tr>";
+                                }
                             }
                             ?>
